Finished a basic achievement system. Made various changes to the achievement classes in order to allow chaining of creating achievements. Also added the ability to add and configure the background/font of the achievement popup as well as manage its display duration. Created a new 'AchievementPopup' class which holds the styling and other default settings for the popup.

+++ marsclicker/Assets/Achievements/AchievementTrigger.cs
@@ -4,11 +4,11 @@ using System.Collections;
 public class AchievementTrigger {
 
     public string TriggerName { get; private set; }
-    public int TriggerActivationValue { get; set; }
-    public int CurrentValue { get; set; }
+    public float TriggerActivationValue { get; set; }
+    public float CurrentValue { get; set; }
 
 
-    public AchievementTrigger(string name, int activationValue)
+    public AchievementTrigger(string name, float activationValue)
     {
         TriggerName = name;
         TriggerActivationValue = activationValue;
